Add file-storage scrubber tags and fix issues

+++ backend/typescript/services/implementations/entityService.ts
@@ -1,5 +1,7 @@
 // mongodb {
+// file-storage {
 import { v4 as uuidv4 } from "uuid";
+// } file-storage
 
 import MgEntity, { Entity } from "../../models/entity.model";
 import {
@@ -7,17 +9,21 @@ import {
   EntityRequestDTO,
   EntityResponseDTO,
 } from "../interfaces/IEntityService";
+// file-storage {
 import IFileStorageService from "../interfaces/fileStorageService";
+// } file-storage
 import logger from "../../utilities/logger";
 
 const Logger = logger(__filename);
 
 class EntityService implements IEntityService {
+  // file-storage {
   storageService: IFileStorageService;
 
   constructor(storageService: IFileStorageService) {
     this.storageService = storageService;
   }
+  // } file-storage
 
   /* eslint-disable class-methods-use-this */
   async getEntity(id: string): Promise<EntityResponseDTO> {
@@ -39,7 +45,9 @@ class EntityService implements IEntityService {
       enumField: entity.enumField,
       stringArrayField: entity.stringArrayField,
       boolField: entity.boolField,
+      // file-storage {
       fileName: entity.fileName,
+      // } file-storage
     };
   }
 
@@ -53,7 +61,9 @@ class EntityService implements IEntityService {
         enumField: entity.enumField,
         stringArrayField: entity.stringArrayField,
         boolField: entity.boolField,
+        // file-storage {
         fileName: entity.fileName,
+        // } file-storage
       }));
     } catch (error) {
       Logger.error(`Failed to get entities. Reason = ${error.message}`);
@@ -63,8 +73,11 @@ class EntityService implements IEntityService {
 
   async createEntity(entity: EntityRequestDTO): Promise<EntityResponseDTO> {
     let newEntity: Entity | null;
+    // file-storage {
     const fileName = entity.filePath ? uuidv4() : "";
+    // } file-storage
     try {
+      // file-storage {
       if (entity.filePath) {
         await this.storageService.createFile(
           fileName,
@@ -73,6 +86,10 @@ class EntityService implements IEntityService {
         );
       }
       newEntity = await MgEntity.create({ ...entity, fileName });
+      // } file-storage
+      // no-file-storage {
+      newEntity = await MgEntity.create(entity);
+      // } no-file-storage
     } catch (error) {
       Logger.error(`Failed to create entity. Reason = ${error.message}`);
       throw error;
@@ -84,7 +101,9 @@ class EntityService implements IEntityService {
       enumField: newEntity.enumField,
       stringArrayField: newEntity.stringArrayField,
       boolField: newEntity.boolField,
+      // file-storage {
       fileName,
+      // } file-storage
     };
   }
 
@@ -93,8 +112,11 @@ class EntityService implements IEntityService {
     entity: EntityRequestDTO,
   ): Promise<EntityResponseDTO | null> {
     let updatedEntity: Entity | null;
+    // file-storage {
     let fileName = "";
+    // } file-storage
     try {
+      // file-storage {
       const currentEntity = await MgEntity.findById(id, "fileName");
       const currentFileName = currentEntity?.fileName;
       if (entity.filePath) {
@@ -123,6 +145,13 @@ class EntityService implements IEntityService {
           runValidators: true,
         },
       );
+      // } file-storage
+      // no-file-storage {
+      updatedEntity = await MgEntity.findByIdAndUpdate(id, entity, {
+        new: true,
+        runValidators: true,
+      });
+      // } no-file-storage
       if (!updatedEntity) {
         throw new Error(`Entity id ${id} not found`);
       }
@@ -137,7 +166,9 @@ class EntityService implements IEntityService {
       enumField: updatedEntity.enumField,
       stringArrayField: updatedEntity.stringArrayField,
       boolField: updatedEntity.boolField,
+      // file-storage {
       fileName,
+      // } file-storage
     };
   }
 
@@ -147,9 +178,11 @@ class EntityService implements IEntityService {
       if (!deletedEntity) {
         throw new Error(`Entity id ${id} not found`);
       }
+      // file-storage {
       if (deletedEntity.fileName) {
         await this.storageService.deleteFile(deletedEntity.fileName);
       }
+      // } file-storage
     } catch (error) {
       Logger.error(`Failed to delete entity. Reason = ${error.message}`);
       throw error;
@@ -161,7 +194,9 @@ export default EntityService;
 
 // } mongodb
 // postgresql {
+// file-storage {
 import { v4 as uuidv4 } from "uuid";
+// } file-storage
 
 import PgEntity from "../../models/entity.model";
 import {
@@ -169,17 +204,21 @@ import {
   EntityRequestDTO,
   EntityResponseDTO,
 } from "../interfaces/IEntityService";
+// file-storage {
 import IFileStorageService from "../interfaces/fileStorageService";
+// } file-storage
 import logger from "../../utilities/logger";
 
 const Logger = logger(__filename);
 
 class EntityService implements IEntityService {
+  // file-storage {
   storageService: IFileStorageService;
 
   constructor(storageService: IFileStorageService) {
     this.storageService = storageService;
   }
+  // } file-storage
 
   /* eslint-disable class-methods-use-this */
   async getEntity(id: string): Promise<EntityResponseDTO> {
@@ -201,7 +240,9 @@ class EntityService implements IEntityService {
       enumField: entity.enum_field,
       stringArrayField: entity.string_array_field,
       boolField: entity.bool_field,
+      // file-storage {
       fileName: entity.file_name,
+      // } file-storage
     };
   }
 
@@ -215,7 +256,9 @@ class EntityService implements IEntityService {
         enumField: entity.enum_field,
         stringArrayField: entity.string_array_field,
         boolField: entity.bool_field,
+        // file-storage {
         fileName: entity.file_name,
+        // } file-storage
       }));
     } catch (error) {
       Logger.error(`Failed to get entities. Reason = ${error.message}`);
@@ -225,8 +268,11 @@ class EntityService implements IEntityService {
 
   async createEntity(entity: EntityRequestDTO): Promise<EntityResponseDTO> {
     let newEntity: PgEntity | null;
+    // file-storage {
     const fileName = entity.filePath ? uuidv4() : "";
+    // } file-storage
     try {
+      // file-storage {
       if (entity.filePath) {
         await this.storageService.createFile(
           fileName,
@@ -234,13 +280,16 @@ class EntityService implements IEntityService {
           entity.fileContentType,
         );
       }
+      // } file-storage
       newEntity = await PgEntity.create({
         string_field: entity.stringField,
         int_field: entity.intField,
         enum_field: entity.enumField,
         string_array_field: entity.stringArrayField,
         bool_field: entity.boolField,
+        // file-storage {
         file_name: fileName,
+        // } file-storage
       });
     } catch (error) {
       Logger.error(`Failed to create entity. Reason = ${error.message}`);
@@ -253,7 +302,9 @@ class EntityService implements IEntityService {
       enumField: newEntity.enum_field,
       stringArrayField: newEntity.string_array_field,
       boolField: newEntity.bool_field,
+      // file-storage {
       fileName,
+      // } file-storage
     };
   }
 
@@ -263,8 +314,11 @@ class EntityService implements IEntityService {
   ): Promise<EntityResponseDTO | null> {
     let resultingEntity: PgEntity | null;
     let updateResult: [number, PgEntity[]] | null;
+    // file-storage {
     let fileName = "";
+    // } file-storage
     try {
+      // file-storage {
       const currentEntity = await PgEntity.findByPk(id, {
         raw: true,
         attributes: ["file_name"],
@@ -288,6 +342,7 @@ class EntityService implements IEntityService {
       } else if (currentFileName) {
         await this.storageService.deleteFile(currentFileName);
       }
+      // } file-storage
       updateResult = await PgEntity.update(
         {
           string_field: entity.stringField,
@@ -295,7 +350,9 @@ class EntityService implements IEntityService {
           enum_field: entity.enumField,
           string_array_field: entity.stringArrayField,
           bool_field: entity.boolField,
+          // file-storage {
           file_name: fileName,
+          // } file-storage
         },
         { where: { id }, returning: true },
       );
@@ -315,23 +372,36 @@ class EntityService implements IEntityService {
       enumField: resultingEntity.enum_field,
       stringArrayField: resultingEntity.string_array_field,
       boolField: resultingEntity.bool_field,
+      // file-storage {
       fileName,
+      // } file-storage
     };
   }
 
   async deleteEntity(id: string): Promise<void> {
     try {
+      // file-storage {
       const entityToDelete = await PgEntity.findByPk(id, { raw: true });
+      // } file-storage
       const deleteResult: number | null = await PgEntity.destroy({
         where: { id },
       });
 
+      // file-storage {
       if (!entityToDelete || !deleteResult) {
         throw new Error(`Entity id ${id} not found`);
       }
+      // } file-storage
+      // no-file-storage {
+      if (!deleteResult) {
+        throw new Error(`Entity id ${id} not found`);
+      }
+      // } no-file-storage
+      // file-storage {
       if (entityToDelete.file_name) {
         await this.storageService.deleteFile(entityToDelete.file_name);
       }
+      // } file-storage
     } catch (error) {
       Logger.error(`Failed to delete entity. Reason = ${error.message}`);
       throw error;
